core: SnapshotManager - disk devices might not represent engine disks
When using the snapshot manager there is an attempt to remove the disks
that aren't part of the snaposhot.

That's being done by loading the vm disk devices, load the disk for each
device and remove it if needed.

If the device represented disk isn't a disk known by the engine, do not
remove it (as being done for disks that aren't snappable/snapshot
disks).

Note - This method could be refactored to load the vm disks
directly instead of the devices and then the disks. In this patch
the null check is be added, while further refactor would be done in
follow up changes (the device are synced multiple times, the vm disks
could be loaded instead of the devices..etc).

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1032807

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/snapshots/SnapshotsManager.java
@@ -345,7 +345,13 @@ public class SnapshotsManager {
         if (!vmDevice.getDevice().equals(VmDeviceType.DISK.getName())) {
             return true;
         }
-        return vmDevice.getSnapshotId() == null && getDiskDao().get(vmDevice.getDeviceId()).isAllowSnapshot();
+
+        if (vmDevice.getSnapshotId() == null) {
+           Disk disk = getDiskDao().get(vmDevice.getDeviceId());
+           return disk != null && disk.isAllowSnapshot();
+        }
+
+        return false;
     }
 
     /**
@@ -513,7 +519,7 @@ public class SnapshotsManager {
                 vmId, VmDeviceGeneralType.DISK, VmDeviceType.DISK.getName())) {
             if (!diskIdsFromSnapshot.contains(vmDevice.getDeviceId()) && vmDevice.getSnapshotId() == null) {
                 Disk disk = getDiskDao().get(vmDevice.getDeviceId());
-                if (disk.isAllowSnapshot()) {
+                if (disk != null && disk.isAllowSnapshot()) {
                     getBaseDiskDao().remove(vmDevice.getDeviceId());
                     getVmDeviceDao().remove(vmDevice.getId());
                 }
